работает регистрация через фбук, добавление ролей VIEWER при регистрации

+++ webscada/api/src/main/java/webscada/api/dao/IRoleJPADao.java
@@ -3,7 +3,7 @@ package webscada.api.dao;
 import org.springframework.data.jpa.repository.JpaRepository;
 import webscada.entity.Role;
 
-public interface IRoleJPADao extends JpaRepository<Role, Long> {
+public interface IRoleJPADao extends JpaRepository<Role, Integer> {
 
 	Role findByRoleName(String roleName);
 }
